Current best version of study 4
this is all I managed to pull off for now. I will submit it in order to minimize the loss of my grade. I think that this still fulfills the requirements albeit in a disappointing way. Im having trouble with the hit box mechanics and I need to figure that out before I do anything else. Ill get back to this as soon as I can, but for now I have much more pressing assignments and learning that I need to do.